The bog idea
I realise that humans need transport and the world doesn't. The carbon footprint that we are leaving is a problem that protesting won't fix. I dream that a platform were wet can recycle already existing cars converting them to electric is the way if the future. Government funding and world sponsoring and investing can make a difference, this is possible and i think with the right marketing we could make a huge step in the right direction. Building electric cars is a great start but unfortunately it isn't recycling and the manufacturing process is still hard on the environment. There are enough preexisting cars on this planet to be restored and converted to electric. Let's make this happen.